+++ HISTORY.md
@@ -9,10 +9,16 @@
     * new animated and still symbols for interlocking gears - used for process extensions
 * **Notable Changes:**
     * changed small gear symbol to 6 spikes (from 8)
+    * MQTT extension update, thanks, Simon and Xavier!
 * **Notable Fixes:**
     * fixed display of continuations capturing variadic inputs
 * **Translation Updates:**
     * German
+
+### 2026-08-03
+* symbols: tweaked gear symbols
+* mqtt extension update, thanks, Simon and Xavier!
+* threads, objects, tables: refactored Process GUI
 
 ### 2026-08-02
 * blocks, byob: new special 'process' input slot type